Move types to dedicated file
